Внешний драйвер.
Закончил логику переход к тестам.

+++ Universal_Driver_soft/UniversalDriver_v2.1_extDriver/librarys/motor.cpp
@@ -26,7 +26,7 @@ void base_motor::SetStepMode(step stepmode){
    StepMode = stepmode;
 }
 
-void base_motor::SetAcceleration(unsigned int accel){
+void base_motor::SetAcceleration(uint16_t accel){
    if(accel <= 100){
       Acceleration = accel;
    }
